Initial Setup
Setup using CLion for MacOS Tahoe. Added a working MakeLists.txt with C++20, and automatic macOS Detection. If anyone is using a different OS it shouldn't be hard to integrate that as well. I integrated SDL2 via homebrew for the graphics and input handling. Added a .gitignore and updated the readme and did some simple verification. Everything compiles and gives a brief executable.